Modificado constructor de personaje e interface objetivo, y los sucesivos cambios que esto implica

+++ tp3/src/grupo9/algo3man/modelo/Objetivo.java
@@ -3,5 +3,5 @@ package grupo9.algo3man.modelo;
 import java.awt.Point;
 
 public interface Objetivo {
-	public int getDistancia(Point posicion);
+	public double getDistancia(Point posicion);
 }
